migrated agent species ids to int from ushort, as to avoid overlapping with real ids when using custom ids

+++ GW2EIParser/Builders/JsonModels/JsonActors/JsonNPC.cs
@@ -13,7 +13,7 @@ namespace GW2EIParser.Builders.JsonModels
         /// <summary>
         /// Game ID of the target
         /// </summary>
-        public ushort Id { get; }
+        public int Id { get; }
         /// <summary>
         /// Total health of the target
         /// </summary>
