+++ GW2EIParser/EIData/Mechanics/MechanicTypes/EnemyBuffRemoveMechanic.cs
@@ -26,7 +26,7 @@ namespace GW2EIParser.EIData
             IsEnemyMechanic = true;
         }
 
-        public override void CheckMechanic(ParsedLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<ushort, AbstractSingleActor> regroupedMobs)
+        public override void CheckMechanic(ParsedLog log, Dictionary<Mechanic, List<MechanicEvent>> mechanicLogs, Dictionary<int, AbstractSingleActor> regroupedMobs)
         {
             foreach (AbstractBuffEvent c in log.CombatData.GetBuffData(SkillId))
             {
